+++ src/checkTestReport.ts
@@ -1,0 +1,3 @@
+export function checkTestReport(_configFile: string) {
+  console.log('Checking test report'); // todo: implement this.
+}
