+++ src/utilities/markdownUtils.ts
@@ -53,7 +53,3 @@ export const buildMarkdownTable = (headers: string[], alignment: Alignment[], da
 };
 
 export const markdownTitle = (title: string, level: number = 1) => [`${'#'.repeat(level)} ${title}`, ''];
-
-export const percentage = (value: number, total: number) => value ? `${((value / total) * 100).toFixed(1)}%` : '';
-
-export const extractFilename = (filename: string) => filename.replace(/\\/g, '/').split('/').at(-1)!;
