+++ test/checkTestReport.test.ts
@@ -1,0 +1,5 @@
+import { checkTestReport } from '../src/checkTestReport.ts';
+
+Deno.test('checkTestReport', () => {
+  checkTestReport('test/testReport.yaml');
+});
